tools/requirements.txt
# Requirements

argparse
pillow
numpy
uuid
datetime
dms2dec
configparser
geographiclib
pyshtools